image object pages are displaying mostly as desired at this point
some blocks still need to be made and/or placed, but this seems like a good checkpoint.

I'd like to make better commit messages, but this whole process of grabbing configs from the ir environment and twisting them to our purpose here is a little hectic and hard to track, so I'm sticking to just committing at mostly-stable points.