workaround for placing a block in a View
as per https://www.drupal.org/project/drupal/issues/2565081
"the block ID only exists as long a the the block is configured to be visible in a region. When the block is deleted from the region, the View breaks."
This workaround of placing a block in a non-rendered region is already being used for facet blocks, so just going with that implementation for this as well.